Backend/requirements.txt
fastapi
uvicorn
mongoose
python-dotenv
sqlalchemy
psycopg2
langchain
langchain-community
langgraph
langchain-groq
serpapi